app/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
Werkzeug==3.0.1
